+++ CHANGELOG.rst
@@ -18,9 +18,17 @@ Unreleased
 
 * nothing unreleased
 
-[8.0.7] - 2026-05-06
+[8.0.9] - 2026-05-07
 ---------------------
 * feat: mark unenrolled courses and filter out audit courses in admin enrollment viewset
+
+[8.0.8] - 2026-05-06
+---------------------
+* feat: extra logging for enterprise email association pipeline step
+
+[8.0.7] - 2026-05-06
+---------------------
+* feat: add pluggable override for course home progress username obfuscation
 
 [8.0.6] - 2026-05-05
 ---------------------